Looking at this diff, I need to analyze what changed across multiple files.
The changes are:
1. `config/mod.rs`: `999999` → `999_999` (numeric literal formatting with underscore)
2. `files/llm_output_extraction.rs`: `r#"..."#` → `r"..."` (raw string literal type)
3. `logger/progress.rs`: Multiple type changes `u32` → `usize`, removal of redundant casts, and simplification of arithmetic
4. `pipeline/tests.rs`: String formatting changes from format args to direct interpolation, and use of raw string literals
5. `prompts/mod.rs`: String formatting changes from format args to direct interpolation

These are all **style** changes - they improve code readability and consistency through:
- Using underscore separators in large numeric literals
- Using simpler raw string literals where escape sequences aren't needed
- Using `usize` consistently instead of `u32` for size-related values
- Removing redundant type casts
- Using modern Rust string interpolation syntax

No behavior changes - just code style improvements.

style: improve code formatting and consistency

- Use underscore separators in large numeric literals (999_999)
- Use raw string literals (r"...") instead of raw bracketed strings where escapes aren't needed
- Change bar_width from u32 to usize for consistency with size operations
- Remove redundant casts and simplify arithmetic expressions
- Use modern string interpolation syntax ({var}) instead of format args

+++ ralph-workflow/src/logger/progress.rs
@@ -28,16 +28,16 @@ pub fn print_progress(current: u32, total: u32, label: &str) {
         return;
     }
 
-    let bar_width: u32 = 20;
+    let bar_width: usize = 20;
     let pct = (u64::from(current))
         .saturating_mul(100)
         .saturating_div(u64::from(total))
         .min(100) as u32;
     let filled = (u64::from(current))
-        .saturating_mul(u64::from(bar_width))
+        .saturating_mul(bar_width as u64)
         .saturating_div(u64::from(total))
-        .min(u64::from(bar_width)) as usize;
-    let empty = bar_width as usize - filled;
+        .min(bar_width as u64) as usize;
+    let empty = bar_width - filled;
 
     let bar: String = "█".repeat(filled) + &"░".repeat(empty);
 
@@ -64,16 +64,16 @@ mod tests {
         if total == 0 {
             return (0, String::new());
         }
-        let bar_width: u32 = 20;
+        let bar_width: usize = 20;
         let pct = (u64::from(current))
             .saturating_mul(100)
             .saturating_div(u64::from(total))
             .min(100) as u32;
         let filled = (u64::from(current))
-            .saturating_mul(u64::from(bar_width))
+            .saturating_mul(bar_width as u64)
             .saturating_div(u64::from(total))
-            .min(u64::from(bar_width)) as usize;
-        let empty = bar_width as usize - filled;
+            .min(bar_width as u64) as usize;
+        let empty = bar_width - filled;
         let bar: String = "█".repeat(filled) + &"░".repeat(empty);
         (pct, bar)
     }
